+++ cpp/pybind/geometry/trianglemesh.cpp
@@ -110,7 +110,8 @@ void pybind_trianglemesh_definitions(py::module &m) {
                  ":math:`v_o = v_i x strength (v_i * |N| - \\sum_{n \\in N} "
                  "v_n)`",
                  "number_of_iterations"_a = 1, "strength"_a = 1,
-                 "filter_scope"_a = MeshBase::FilterScope::All)
+                 py::arg_v("filter_scope", MeshBase::FilterScope::All,
+                           "FilterScope.All"))
             .def("filter_smooth_simple", &TriangleMesh::FilterSmoothSimple,
                  "Function to smooth triangle mesh with simple neighbour "
                  "average. :math:`v_o = \\frac{v_i + \\sum_{n \\in N} "
@@ -118,7 +119,8 @@ void pybind_trianglemesh_definitions(py::module &m) {
                  ":math:`v_o` the output value, and :math:`N` is the set of "
                  "adjacent neighbours.",
                  "number_of_iterations"_a = 1,
-                 "filter_scope"_a = MeshBase::FilterScope::All)
+                 py::arg_v("filter_scope", MeshBase::FilterScope::All,
+                           "FilterScope.All"))
             .def("filter_smooth_laplacian",
                  &TriangleMesh::FilterSmoothLaplacian,
                  "Function to smooth triangle mesh using Laplacian. :math:`v_o "
@@ -129,7 +131,8 @@ void pybind_trianglemesh_definitions(py::module &m) {
                  "inverse distance (closer neighbours have higher weight), and "
                  "lambda_filter is the smoothing parameter.",
                  "number_of_iterations"_a = 1, "lambda_filter"_a = 0.5,
-                 "filter_scope"_a = MeshBase::FilterScope::All)
+                 py::arg_v("filter_scope", MeshBase::FilterScope::All,
+                           "FilterScope.All"))
             .def("filter_smooth_taubin", &TriangleMesh::FilterSmoothTaubin,
                  "Function to smooth triangle mesh using method of Taubin, "
                  "\"Curve and Surface Smoothing Without Shrinkage\", 1995. "
@@ -139,7 +142,9 @@ void pybind_trianglemesh_definitions(py::module &m) {
                  "parameter mu as smoothing parameter. This method avoids "
                  "shrinkage of the triangle mesh.",
                  "number_of_iterations"_a = 1, "lambda_filter"_a = 0.5,
-                 "mu"_a = -0.53, "filter_scope"_a = MeshBase::FilterScope::All)
+                 "mu"_a = -0.53,
+                 py::arg_v("filter_scope", MeshBase::FilterScope::All,
+                           "FilterScope.All"))
             .def("has_vertices", &TriangleMesh::HasVertices,
                  "Returns ``True`` if the mesh contains vertices.")
             .def("has_triangles", &TriangleMesh::HasTriangles,
@@ -250,7 +255,9 @@ void pybind_trianglemesh_definitions(py::module &m) {
                  &TriangleMesh::SimplifyVertexClustering,
                  "Function to simplify mesh using vertex clustering.",
                  "voxel_size"_a,
-                 "contraction"_a = MeshBase::SimplificationContraction::Average)
+                 py::arg_v("contraction",
+                           MeshBase::SimplificationContraction::Average,
+                           "SimplificationContraction.Average"))
             .def("simplify_quadric_decimation",
                  &TriangleMesh::SimplifyQuadricDecimation,
                  "Function to simplify mesh using Quadric Error Metric "
@@ -298,7 +305,9 @@ void pybind_trianglemesh_definitions(py::module &m) {
                  "'As-Rigid-As-Possible Surface Modeling', 2007",
                  "constraint_vertex_indices"_a, "constraint_vertex_positions"_a,
                  "max_iter"_a,
-                 "energy"_a = MeshBase::DeformAsRigidAsPossibleEnergy::Spokes,
+                 py::arg_v("energy",
+                           MeshBase::DeformAsRigidAsPossibleEnergy::Spokes,
+                           "DeformAsRigidAsPossibleEnergy.Spokes"),
                  "smoothed_alpha"_a = 0.01)
             .def_static(
                     "create_from_point_cloud_alpha_shape",