+++ cpp/pybind/pipelines/registration/feature.cpp
@@ -19,6 +19,10 @@ void pybind_feature_declarations(py::module &m_registration) {
     py::class_<Feature, std::shared_ptr<Feature>> feature(
             m_registration, "Feature",
             "Class to store featrues for registration.");
+    m_registration.attr("m") =
+            py::module_::import("typing").attr("TypeVar")("m");
+    m_registration.attr("n") =
+            py::module_::import("typing").attr("TypeVar")("n");
 }
 void pybind_feature_definitions(py::module &m_registration) {
     // open3d.registration.Feature
@@ -31,6 +35,10 @@ void pybind_feature_definitions(py::module &m_registration) {
             .def("dimension", &Feature::Dimension,
                  "Returns feature dimensions per point.")
             .def("num", &Feature::Num, "Returns number of points.")
+            .def("select_by_index", &Feature::SelectByIndex,
+                 "Function to select features from input Feature group into "
+                 "output Feature group.",
+                 "indices"_a, "invert"_a = false)
             .def_readwrite("data", &Feature::data_,
                            "``dim x n`` float64 numpy array: Data buffer "
                            "storing features.")
@@ -48,6 +56,11 @@ void pybind_feature_definitions(py::module &m_registration) {
     docstring::ClassMethodDocInject(m_registration, "Feature", "resize",
                                     {{"dim", "Feature dimension per point."},
                                      {"n", "Number of points."}});
+    docstring::ClassMethodDocInject(
+            m_registration, "Feature", "select_by_index",
+            {{"indices", "Indices of features to be selected."},
+             {"invert",
+              "Set to ``True`` to invert the selection of indices."}});
     m_registration.def("compute_fpfh_feature", &ComputeFPFHFeature,
                        "Function to compute FPFH feature for a point cloud",
                        "input"_a, "search_param"_a);